pyproject.toml
[tool.poetry]
name = "proxy6-middleware"
version = "0.1.0"
description = ""
authors = ["Mailf <mailf.channel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
django-cors-headers = "^4.7.0"
httpx = "^0.28.1"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.1"
django-filter = "^25.1"
dj-database-url = "^3.0.1"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
